+++ NEWS.md
@@ -4,6 +4,15 @@ This page describes the most important changes in `Neutrals`. The format is base
 a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic
 Versioning](https://semver.org/spec).
 
+## Unreleased
+
+### Added
+
+- Ranges can be constructed with neutral numbers specified as the start, step, and/or stop
+  parameters of the range. `𝟙:stop` is identical to `Base.OneTo(stop)` if `stop` is a
+  non-neutral integer or is `𝟙` and is identical to `Base.OneTo(Int(stop))` otherwise.
+  `start:𝟙:stop` identical to `start:stop` whatever, `start` and `stop`.
+
 
 ## Version 0.3.0 (2025/05/26)